util functions in util folder
